PersistenceSpec small refactor
